--- frontend/src/app/board/page.tsx.backup
@@ -1,22 +1,0 @@
-'use client';
-
-import { useEffect } from 'react';
-import { useRouter } from 'next/navigation';
-
-export default function BoardPage() {
-  const router = useRouter();
-
-  useEffect(() => {
-    // Redirect to dashboard by default
-    router.replace('/board/dashboard');
-  }, [router]);
-
-  return (
-    <div className="min-h-screen bg-gradient-to-b from-zinc-50 via-white to-zinc-50 flex items-center justify-center">
-      <div className="text-center">
-        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-zinc-600 mx-auto"></div>
-        <p className="mt-4 text-zinc-600">Loading...</p>
-      </div>
-    </div>
-  );
-}
